requirements.txt
fastapi
google-api-core
google-auth
google-cloud-core
google-cloud-speech
google-cloud-storage
python-dotenv
requests
twilio
uvicorn